requirements.txt
requests
google-genai
pyyaml
click
